requirements.txt
pip==20.0.2
fastapi==0.47.1
uvicorn==0.11.2
PyYAML==5.3
email-validator==1.0.5
pytest==5.3.4
pycodestyle==2.5.0
boto3==1.10.32
moto==1.3.14
